// frontend/packages/core/src/connection/types.ts
export type Credentials = RequestInit['credentials'];
export type Endpoint = string;
export type FetchRetryOptions = {
  retries?: number;
  retryDelay?:
    | number
    | ((attempt: number, error: Error, response: Response) => number);
  retryOn?:
    | number[]
    | ((attempt: number, error: Error, response: Response) => boolean);
};
export type Headers = { [k: string]: string };
export type Host = string;
// ...
export type Mode = RequestInit['mode'];
export type Redirect = RequestInit['redirect'];
export type ClientTimeout = number | undefined;
// ...
export type CsrfToken = string;
export type CsrfPromise = Promise<string | undefined>;
export type Protocol = 'http:' | 'https:';

export interface ClientConfig {
  host?: Host;
  protocol?: Protocol;
  appRoot?: string;
  credentials?: Credentials;
  csrfToken?: CsrfToken;
  guestToken?: string;
  guestTokenHeaderName?: string;
  fetchRetryOptions?: FetchRetryOptions;
  headers?: Headers;
  mode?: Mode;
  timeout?: ClientTimeout;
  unauthorizedHandler?: () => void;
}

export interface ZobiClientInterface extends Pick<
  ZobiClientClass,
  | 'delete'
  | 'get'
  | 'post'
  | 'postForm'
  | 'put'
  | 'request'
  | 'init'
  | 'isAuthenticated'
  | 'reAuthenticate'
  | 'getGuestToken'
  | 'getUrl'
> {
  configure: (config?: ClientConfig) => ZobiClientInterface;
  reset: () => void;
  getCSRFToken: () => CsrfPromise;
}